requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
google-generativeai
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]